Testing
I realized that you don't have to always get 30 eggs per loop, because any blank spaces left in the boxes from not having 30 will be filled in on the second round. I still have to figure out what happens if you try to take an egg with a full inventory, and script in that change.