Move commands and subscripts to subscript.py
Commands and Subscripts are now in subscripts.py. It is included in Make_Script.py. There is a function called get in subscripts that returns the subscripts as lists when called with the parameter of the name of the variable in quotes, as they're all in a dictionary that way.